+++ server/services/TagService.js
@@ -1,4 +1,4 @@
-const Session = require('../database/Session');
+const Session = require('../infra/database/Session');
 const Tag = require('../models/Tag');
 
 class TagService {
