+++ server/services/TreeTagService.js
@@ -1,4 +1,4 @@
-const Session = require('../database/Session');
+const Session = require('../infra/database/Session');
 const TreeTag = require('../models/TreeTag');
 
 class TreeTagService {
